+++ README.md
@@ -10,7 +10,22 @@
 
 A Java client for [SCION](https://scion.org).
 
+### WARNING
 This client can directly connect to SCION **without dispatcher**.
+
+Currently (January 2024), the SCION system uses a "dispatcher" (a process that runs on endhosts,
+listens on a fixed port (30041) and forwards any incoming SCION packets, after stripping the SCION 
+header, to local application).
+
+This Java client cannot be used with a dispatcher.
+The Java client can be used in one of the following ways:
+- You can use the client stand-alone (without local SCION installation),
+  however it must listen on port 30041 for incoming SCION packets because
+  SCION routers currently will forward data only to that port. 
+- If you need a local SCION installation (Go implementation),
+  consider using the dispatch-off branch/PR.
+- When you need to run a local system with dispatcher, you can try to use port forwarding
+  to forward incoming data your Java application port. The application port must not be 30041.
 
 ## API
 
@@ -75,7 +90,35 @@ try (DatagramChannel channel = DatagramChannel.open()) {
 }
 ```
 
+### Demos
+
+Some demos can be found in [src/test/java/org/scion/demo](src/test/java/org/scion/demo).
+
+- `DatagramChannel` ping pong [client](src/test/java/org/scion/demo/ScionPingPongChannelClient.java) 
+  and [server](src/test/java/org/scion/demo/ScionPingPongChannelServer.java)
+- [SCMP echo](src/test/java/org/scion/demo/ScmpEchoDemo.java)
+- [SCMP traceroute](src/test/java/org/scion/demo/ScmpTracerouteDemo.java)
+
+
+### General documentation
+
+
+- Reference manual: https://docs.scion.org
+- SCIONLab, a global testbed for SCION applications: https://www.scionlab.org/
+- Awesome SCION, a collection of SCION projects: https://github.com/scionproto/awesome-scion 
+
 ## DatagramChannel
+
+### Demo application - ping pong
+
+There is a simple ping pong client-server application in `src/test/demo`.
+
+It has some hardcoded ports/IP so it works only with the scionlab tiny.topo and only with the dispatcher-free
+version of scionlab: https://github.com/scionproto/scion/pull/4344
+
+The client and server connects directly to the border router (without dispatcher).
+
+The server is located in `1-ff00:0:112` (IP [::1]:44444). The client is located in `1-ff00:0:110`.
 
 ### Options
 
@@ -97,23 +140,7 @@ Options are defined in `ScionSocketOptions`, see javadoc for details.
   Solution: always use the latest path returned by send, e.g. `path = send(buffer, path)`.
 
 - **Using expired path (server).** When using `send(buffer, path)` with an expired `ResponsePath`, the channel will
-  simple send it anyway (could just drop it) TODO
-  -> Callback?
-  - TODO request new path a few seconds in advance on client side?    
-
-
-
-## Demo application - ping pong
-
-There is a simple ping pong client-server application in `src/test/demo`.
-
-It has some hardcoded ports/IP so it works only with the scionlab tiny.topo and only with the dispatcher-free
-version of scionlab: https://github.com/scionproto/scion/pull/4344
-
-The client and server connects directly to the border router (without dispatcher).
-
-The server is located in `1-ff00:0:112` (IP [::1]:44444). The client is located in `1-ff00:0:110`.
-
+  simple send it anyway.
 
 ## Configuration
 
@@ -147,6 +174,16 @@ the other options are skipped if no property or environment variable is defined.
 | Path expiry margin. Before sending a packet a new path is requested if the path is about to expire with X seconds. | `org.scion.pathExpiryMargin` | `SCION_PATH_EXPIRY_MARGIN`  | 10           |
 
 ## FAQ / trouble shooting
+
+### Local testbed (scionproto) does not contain any path
+
+A common problem is that the certificates of the testbed have expired (default validity: 3 days).
+The certificates can be renewed by recreating the network with 
+`./scion.sh topology -c <your_topology_here.topo>`.
+
+### ERROR: "TRC NOT FOUND"
+This error occurs when requesting a path with an ISD/AS code that is not
+known in the network.
 
 ### Cannot find symbol javax.annotation.Generated
